security: cwe-14
sensitive memory is cleared according to the source code, but compiler optimizations leave the memory untouched when it is not read from again, aka dead store removal.

there is 1 remaining memset on L28 in ecc_tests.c that still need to be addressed but because it sets value to 255 we can't use dogecoin_mem_zero.

https://cwe.mitre.org/data/definitions/14.html